根据用户隔离相应资源

+++ src/main/java/com/lizijian/officeauto/Controller/RoleController.java
@@ -3,11 +3,13 @@ package com.lizijian.officeauto.Controller;
 import com.lizijian.officeauto.Service.RoleService;
 import com.lizijian.officeauto.mapper.RoleMapper;
 import com.lizijian.officeauto.pojo.WebApiResult;
+import com.lizijian.officeauto.utils.ResourcesAuthenticateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurationSupport;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 
 @RestController
@@ -17,23 +19,35 @@ public class RoleController {
     @Autowired
     RoleService roleService;
 
+    @Autowired
+    ResourcesAuthenticateUtils resourcesAuthenticateUtils;
+
     @DeleteMapping
-    public WebApiResult deleteUserRole(HttpServletRequest request){
+    public WebApiResult deleteUserRole(HttpServletRequest request, HttpServletResponse response){
         Integer rid = Integer.valueOf(request.getParameter("rid"));
         Integer uid = Integer.valueOf(request.getParameter("uid"));
-        return roleService.deleteUserRole(uid, rid);
+        if (resourcesAuthenticateUtils.assertStuffInAuthenticateResources(request,uid)){
+            return roleService.deleteUserRole(uid, rid);
+        }
+        else {
+            response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+            return null;
+        }
     }
 
     @PostMapping
-    public WebApiResult insertUserRole(Integer uid, Integer rid){
+    public WebApiResult insertUserRole(HttpServletRequest request, HttpServletResponse response, Integer uid, Integer rid){
         if (uid == (null) || rid ==null){
             WebApiResult webApiResult = new WebApiResult();
             webApiResult.isNull();
             webApiResult.setMsg("参数为空!!!");
             return webApiResult;
         }
-        return roleService.insertUserRole(uid, rid);
+        if (resourcesAuthenticateUtils.assertStuffInAuthenticateResources(request,uid)){
+            return roleService.insertUserRole(uid, rid);
+        }
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+        return null;
     }
-
 
 }
